+++ src/main/java/de/teragam/jfxshader/material/internal/es2/ES2ShaderMeshView.java
@@ -1,11 +1,12 @@
 package de.teragam.jfxshader.material.internal.es2;
 
 import com.sun.prism.Material;
-import com.sun.prism.es2.ES2ResourceFactory;
+import com.sun.prism.ResourceFactory;
 import com.sun.prism.impl.BaseMesh;
 import com.sun.prism.impl.Disposer;
 import com.sun.prism.impl.ps.BaseShaderContext;
 
+import de.teragam.jfxshader.exception.ShaderException;
 import de.teragam.jfxshader.material.internal.ShaderMeshView;
 import de.teragam.jfxshader.util.Reflect;
 
@@ -20,7 +21,10 @@ public class ES2ShaderMeshView extends ShaderMeshView {
         this.es2Context = es2Context;
     }
 
-    public static ES2ShaderMeshView create(ES2ResourceFactory rf, BaseMesh mesh) {
+    public static ES2ShaderMeshView create(ResourceFactory rf, BaseMesh mesh) {
+        if (!Reflect.resolveClass("com.sun.prism.es2.ES2ResourceFactory").isInstance(rf)) {
+            throw new ShaderException("Factory is not a ES2ResourceFactory");
+        }
         final BaseShaderContext es2Context = Reflect.on(rf.getClass()).getFieldValue("context", rf);
         final long nativeHandle = Reflect.on(es2Context.getClass()).<Long>method("createES2MeshView").invoke(es2Context, mesh);
         return new ES2ShaderMeshView(nativeHandle, mesh, new ES2MeshViewDisposerRecord(es2Context, nativeHandle), es2Context);