+++ src/main/java/module-info.java
@@ -1,0 +1,14 @@
+module de.teragam.jfxshader {
+    requires javafx.base;
+    requires javafx.graphics;
+
+    exports de.teragam.jfxshader;
+    exports de.teragam.jfxshader.effect;
+    exports de.teragam.jfxshader.material;
+    exports de.teragam.jfxshader.exception;
+    exports de.teragam.jfxshader.renderstate;
+    exports de.teragam.jfxshader.samples.effects.pixelate;
+    exports de.teragam.jfxshader.samples.effects.blendshapes;
+    exports de.teragam.jfxshader.samples.effects.zoomradialblur;
+    exports de.teragam.jfxshader.samples.materials.fresnel;
+}